Add pre-capture hook function

+++ indi_allsky/image.py
@@ -2105,26 +2105,39 @@ class ImageWorker(Process):
 
         save_hook_timeout = self.config.get('IMAGE_SAVE_HOOK_TIMEOUT', 5)
 
-        while self._hookPidRunning():
+        while self._processRunning(self.image_save_hook_process):
             now_time = time.time()
-            if now_time - self.image_save_hook_process_start > save_hook_timeout:
-                logger.error('Image pre-save script exceeded runtime')
-                break
-
-            time.sleep(0.1)
-
-
-        for _ in range(5):
-            if self._hookPidRunning():
-                self.image_save_hook_process.terminate()
-                time.sleep(0.5)
+            if now_time - self.image_save_hook_process_start < save_hook_timeout:
+                time.sleep(0.1)
                 continue
-            else:
-                break
 
-        else:
-            logger.error('Killing image pre-save script')
-            self.image_save_hook_process.kill()
+
+            logger.error('Image pre-save script exceeded runtime')
+
+            for _ in range(5):
+                if not self._processRunning(self.image_save_hook_process):
+                    break
+
+                self.image_save_hook_process.terminate()
+                time.sleep(0.25)
+                continue
+
+
+            if self._processRunning(self.image_save_hook_process):
+                logger.error('Killing image pre-save script')
+                self.image_save_hook_process.kill()
+                self.image_save_hook_process.poll()  # close out process
+
+
+            try:
+                self.pre_hook_datajson_name_p.unlink()
+            except FileNotFoundError:
+                pass
+            except PermissionError as e:
+                logger.error('Unable to delete temp file: %s', str(e))
+
+
+            return {}
 
 
         stdout, stderr = self.image_save_hook_process.communicate()
@@ -2184,26 +2197,30 @@ class ImageWorker(Process):
 
         save_hook_timeout = self.config.get('IMAGE_SAVE_HOOK_TIMEOUT', 5)
 
-        while self._hookPidRunning():
+        while self._processRunning(self.image_save_hook_process):
             now_time = time.time()
-            if now_time - self.image_save_hook_process_start > save_hook_timeout:
-                logger.error('Image post-save script exceeded runtime')
-                break
-
-            time.sleep(0.1)
-
-
-        for _ in range(5):
-            if self._hookPidRunning():
-                self.image_save_hook_process.terminate()
-                time.sleep(0.5)
+            if now_time - self.image_save_hook_process_start < save_hook_timeout:
+                time.sleep(0.1)
                 continue
-            else:
-                break
 
-        else:
-            logger.error('Killing image post-save script')
-            self.image_save_hook_process.kill()
+
+            logger.error('Image post-save script exceeded runtime')
+
+            for _ in range(5):
+                if not self._processRunning(self.image_save_hook_process):
+                    break
+
+                self.image_save_hook_process.terminate()
+                time.sleep(0.25)
+                continue
+
+
+            if self._processRunning(self.image_save_hook_process):
+                logger.error('Killing image post-save script')
+                self.image_save_hook_process.kill()
+                self.image_save_hook_process.poll()  # close out process
+
+            return
 
 
         stdout, stderr = self.image_save_hook_process.communicate()
@@ -2219,12 +2236,12 @@ class ImageWorker(Process):
         self.image_save_hook_process = None
 
 
-    def _hookPidRunning(self):
-        if not self.image_save_hook_process:
+    def _processRunning(self, process):
+        if not process:
             return False
 
         # poll returns None when process is active, rc (normally 0) when finished
-        poll = self.image_save_hook_process.poll()
+        poll = process.poll()
         if isinstance(poll, type(None)):
             return True
